// React/app/GerAdv_TS/ProObservacoes/Interfaces/interface.ProObservacoes.ts
// 100% auto-generated by Source Genesys WebApi Front & Back Creator
// copyright © 2000-2025 Menphis - Sistemas Inteligentes
// This file is part of the Source Genesys project                     
'use client';

export interface IProObservacoes {
// 202501251
    id: number;
 
	processo: number,
	nome: string,
	observacoes: string,
	data: string,
}

export interface IProObservacoesFormProps {
  id: number;
  onClose: () => void;
  onSuccess: (registro?: any) => void;
  onError: () => void;
}

export interface IProObservacoesIncProps {
    id: number;
    onClose: () => void;
    onSuccess: (registro?: any) => void;
    onError: () => void;
}